# Specialize navigating map[string]interface{}